separated server components
