+++ react-app/src/components/EditServerForm/index.js
@@ -1,0 +1,3 @@
+import EditServerForm from "./EditServerForm";
+
+export default EditServerForm;
