+++ react-app/src/components/Nav/NavBar.js
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { NavLink, useHistory } from "react-router-dom";
 
 import { showModal, setCurrentModal } from "../../store/modal";
-import EditServerForm from "../Server/EditServerForm";
+import EditServerForm from "../EditServerForm";
 import Avatar from "../UserInfo/Avatar";
 import { logout } from "../../store/session";
