+++ react-app/src/components/ServerDeleteButton/index.js
@@ -1,0 +1,3 @@
+import ServerDeleteButton from "./ServerDeleteButton";
+
+export default ServerDeleteButton;
